+++ ubuntu.sh
@@ -4,15 +4,19 @@ apt update -y
 # apt upgrade -y
 
 # python3.6
+if [[ `which python3.6` ]];then
 echo '--------------------install python3.6--------------------'
 apt install -y software-properties-common
 add-apt-repository -y ppa:deadsnakes/ppa
 apt -y update
 apt install -y python3.6
+fi
 
 # pip
+if [[ ! `which pip` ]];then
 echo '--------------------install pip--------------------'
 curl https://bootstrap.pypa.io/get-pip.py | python3.6
+fi
 
 # zsh
 echo '--------------------install zsh--------------------'
@@ -21,12 +25,16 @@ apt install -y zsh
 
 
 # oh-my-zsh
+if [[ ! -e ~/.oh-my-zsh ]];then
 echo '--------------------install oh-my-zsh--------------------'
 sh -c "$(wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
+fi
 
 # zsh-completions
+if [[ ! -e ~/.oh-my-zsh/custom/plugins/zsh-completions ]];then
 cd ~
 git clone https://github.com/zsh-users/zsh-completions ~/.oh-my-zsh/custom/plugins/zsh-completions
+fi
 
 # vim 8
 echo '--------------------install vim8--------------------'
@@ -43,11 +51,13 @@ echo '--------------------install tmux--------------------'
 apt install -y tmux
 
 # oh-my-tmux
+if [[ ! -e ~/.tmux ]];then
 echo '--------------------install oh-my-tmux--------------------'
 cd ~
 git clone https://github.com/gpakosz/.tmux.git
 ln -s -f .tmux/.tmux.conf
 cp .tmux/.tmux.conf.local .
+fi
 
 # not ack2
 echo '--------------------install ag--------------------'
@@ -55,9 +65,11 @@ apt install silversearcher-ag
 
 
 # fzf
+if [[ ! -e ~/.fzf ]];then
 echo '--------------------install fzf--------------------'
 t clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
 ~/.fzf/install --all
+fi
 
 
 echo '--------------------install fasd--------------------'
